Serial fixes
Now works, the bug was that no serial port parameters where set, and be
sure that those params match the emulated port params for it to work.
BTW there was error with readline function so now i read everything and
it works fine. Next step Mairiusz check this lib to use the igi xbee
https://github.com/digidotcom/xbee_ansic_library